Fix child selector issue in `ornament`
`lambdaisland/ornament` ignores the class names generated by `girouette`
and substitutes its own. This causes issues when `girouette` returns not
only a class name but also some additional selector. In this case, the
family of tailwind utilities like `space-x-2` was causing issues because
it uses a common CSS selector pattern of: `.space-x-2 > * + *`.

To fix the issue, I've modified the implementation of several tw
utilities to have one additional layer of nesting in the garden output.

See the changes to the unit tests for an example.