Workaround to address issue in garden
Garden's `&` selector does not behave as expected and this is a
hacky way around the problem.

In garden, the & symbol is used to target the parent element; however,
it does not work for the following selector:
`& * + *`. See: https://github.com/noprompt/garden/issues/187

By removing all the spaces in the selector: `&>*+*`, it does work which
is what this commit changes.