✨ Support for using `px` as default number unit
Tailwind uses it's own number units. So for class name `p-1`, the number
is `1`, and for `p-4` the number is `4`.

This tailwind unit typically equated to `1 unit = 0.25rem`.

Rem values are dependent on the `html` `font-size` value. In complex
situations like using `shadow-roots`, `rem`s no longer provide isolation,
by being dependent on the host css styles.

To avoid this use case, we add support for `px` values such that optionally
the tailwind unit can become `1 unit = 4px`. This is equivalent to
`html { font-size: 16px }`, which is typically what browsers default to.